Removed "comparison between signed and unsigned integer expressions" warnings in diagraphica.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5124 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/attrdiscr.cpp
@@ -170,7 +170,7 @@ void AttrDiscr::moveValue(
         if ( idxFr < idxTo )
         {
             // move all values after idxFr 1 pos up
-            for ( int i = idxFr; i < idxTo; ++i )
+            for ( size_t i = idxFr; i < idxTo; ++i )
             {
                 curValues[i] = curValues[i+1];
                 curValues[i]->setIndex( i );
@@ -182,7 +182,7 @@ void AttrDiscr::moveValue(
         else if ( idxTo < idxFr )
         {
             // move all values before idxFr 1 pos down
-            for ( int i = idxFr; i > idxTo; --i )
+            for ( size_t i = idxFr; i > idxTo; --i )
             {
                 curValues[i] = curValues[i-1];
                 curValues[i]->setIndex( i );
